Rename to avoid duplicate mock
