+++ public/app/plugins/datasource/cloud-monitoring/__mocks__/cloudMonitoringInstanceSettings.ts
